I made a flood fill that can color a singular card given that its white.
