+++ src/RunMe.java
@@ -3,7 +3,7 @@ import core.DisplayWindow;
 public class RunMe {
     public static void main(String[] args) {
         // --== Load an image to filter ==--
-        DisplayWindow.showFor("images/image1.jpg", 800, 600, "DoNothingFilter");
+        DisplayWindow.showFor("images/image4.jpg", 800, 600, "sdfasdsf");
 
         // --== Determine your input interactively with menus ==--
 //        DisplayWindow.getInputInteractively(800,600);
